Improvements
- fix(ui): fix nav bar z-index issue
feat(db): migrate workout duration to decimal
- feat(workouts): update duration handling and dialog display
- test(utils): add tests for pace utilities
- test(utils): add tests for date utilities
- ci: add GitHub Actions workflows

+++ types/workoutValidation.ts
@@ -34,7 +34,7 @@ export const workoutFormSchema = z.object({
     .string()
     .transform((val) => {
       if (!val || val === "") return undefined;
-      const num = parseInt(val);
+      const num = parseFloat(val);
       if (isNaN(num) || num <= 0) {
         throw new Error("Duration must be a positive number");
       }
